requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
coverage==7.6.3
dnspython==2.7.0
email_validator==2.2.0
Faker==30.3.0
flake8==7.1.1
flake8-html==0.4.3
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mysqlclient==2.2.4
packaging==24.1
pluggy==1.5.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-ordering==0.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rich==13.9.2
sentry-sdk==2.17.0
setuptools==75.2.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
